+++ lib/core/utils/responsive.dart
@@ -1,0 +1,85 @@
+import 'package:flutter/material.dart';
+
+class Responsive {
+  static const double mobileBreakpoint = 600;
+  static const double tabletBreakpoint = 900;
+  static const double desktopBreakpoint = 1200;
+  static const double largeDesktopBreakpoint = 1800;
+
+  static bool isMobile(BuildContext context) {
+    return MediaQuery.of(context).size.width < mobileBreakpoint;
+  }
+
+  static bool isTablet(BuildContext context) {
+    final width = MediaQuery.of(context).size.width;
+    return width >= mobileBreakpoint && width < tabletBreakpoint;
+  }
+
+  static bool isDesktop(BuildContext context) {
+    final width = MediaQuery.of(context).size.width;
+    return width >= tabletBreakpoint && width < desktopBreakpoint;
+  }
+
+  static bool isLargeDesktop(BuildContext context) {
+    return MediaQuery.of(context).size.width >= desktopBreakpoint;
+  }
+
+  static bool isFoldable(BuildContext context) {
+    final mediaQuery = MediaQuery.of(context);
+    return mediaQuery.size.width >= tabletBreakpoint ||
+        mediaQuery.size.height >= tabletBreakpoint;
+  }
+
+  static double getScreenWidth(BuildContext context) {
+    return MediaQuery.of(context).size.width;
+  }
+
+  static double getScreenHeight(BuildContext context) {
+    return MediaQuery.of(context).size.height;
+  }
+
+  static double getMaxContentWidth(BuildContext context) {
+    final width = getScreenWidth(context);
+    if (width >= largeDesktopBreakpoint) {
+      return 1400;
+    } else if (width >= desktopBreakpoint) {
+      return 1200;
+    } else if (width >= tabletBreakpoint) {
+      return 900;
+    } else {
+      return width;
+    }
+  }
+
+  static EdgeInsets getPadding(BuildContext context) {
+    if (isMobile(context)) {
+      return const EdgeInsets.all(16);
+    } else if (isTablet(context)) {
+      return const EdgeInsets.all(24);
+    } else {
+      return const EdgeInsets.symmetric(horizontal: 48, vertical: 32);
+    }
+  }
+
+  static int getCrossAxisCount(BuildContext context) {
+    if (isMobile(context)) {
+      return 1;
+    } else if (isTablet(context)) {
+      return 2;
+    } else if (isDesktop(context)) {
+      return 3;
+    } else {
+      return 4;
+    }
+  }
+
+  static double getColumnSpacing(BuildContext context) {
+    if (isMobile(context)) {
+      return 16;
+    } else if (isTablet(context)) {
+      return 24;
+    } else {
+      return 32;
+    }
+  }
+}